+++ dango/testing/src/hyperlane.rs
@@ -1,4 +1,130 @@
-mod suite;
-mod validator_set;
+use {
+    crate::{MockValidatorSets, constants::MOCK_HYPERLANE_LOCAL_DOMAIN},
+    anyhow::anyhow,
+    dango_genesis::Contracts,
+    dango_types::{gateway::Domain, warp::TokenMessage},
+    grug_app::{AppError, Db, Indexer, NaiveProposalPreparer, NullIndexer, ProposalPreparer, Vm},
+    grug_db_memory::MemDb,
+    grug_math::Uint128,
+    grug_testing::TestSuite,
+    grug_types::{Addr, Addressable, Coins, Hash256, Signer},
+    grug_vm_rust::RustVm,
+    hyperlane_types::{Addr32, mailbox},
+    std::ops::{Deref, DerefMut},
+};
 
-pub use {suite::*, validator_set::*};
+pub struct HyperlaneTestSuite<DB = MemDb, VM = RustVm, PP = NaiveProposalPreparer, ID = NullIndexer>
+where
+    DB: Db,
+    VM: Vm + Clone + 'static,
+    PP: ProposalPreparer,
+    ID: Indexer,
+    AppError: From<DB::Error> + From<VM::Error> + From<PP::Error>,
+{
+    suite: TestSuite<DB, VM, PP, ID>,
+    validator_sets: MockValidatorSets,
+    mailbox: Addr,
+    warp: Addr,
+}
+
+impl<DB, VM, PP, ID> Deref for HyperlaneTestSuite<DB, VM, PP, ID>
+where
+    DB: Db,
+    VM: Vm + Clone + 'static,
+    PP: ProposalPreparer,
+    ID: Indexer,
+    AppError: From<DB::Error> + From<VM::Error> + From<PP::Error>,
+{
+    type Target = TestSuite<DB, VM, PP, ID>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.suite
+    }
+}
+
+impl<DB, VM, PP, ID> DerefMut for HyperlaneTestSuite<DB, VM, PP, ID>
+where
+    DB: Db,
+    VM: Vm + Clone + 'static,
+    PP: ProposalPreparer,
+    ID: Indexer,
+    AppError: From<DB::Error> + From<VM::Error> + From<PP::Error>,
+{
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.suite
+    }
+}
+
+impl<DB, VM, PP, ID> HyperlaneTestSuite<DB, VM, PP, ID>
+where
+    DB: Db,
+    VM: Vm + Clone + Send + Sync + 'static,
+    PP: ProposalPreparer,
+    ID: Indexer,
+    AppError: From<DB::Error> + From<VM::Error> + From<PP::Error>,
+{
+    pub fn new(
+        suite: TestSuite<DB, VM, PP, ID>,
+        validator_sets: MockValidatorSets,
+        contracts: &Contracts,
+    ) -> Self {
+        Self {
+            suite,
+            validator_sets,
+            mailbox: contracts.hyperlane.mailbox,
+            warp: contracts.warp,
+        }
+    }
+
+    pub async fn receive_warp_transfer<R, A>(
+        &mut self,
+        relayer: &mut (dyn Signer + Send + Sync),
+        origin_domain: Domain,
+        origin_warp: Addr32,
+        recipient: &R,
+        amount: A,
+    ) -> anyhow::Result<Hash256>
+    where
+        R: Addressable,
+        A: Into<Uint128>,
+    {
+        // Mock validator set signs the message.
+        let (message_id, raw_message, raw_metadata) = self
+            .validator_sets
+            .get(origin_domain)
+            .ok_or_else(|| {
+                anyhow!(
+                    "[HyperlaneTestSuite]: no mock validator set found for domain `{origin_domain}`"
+                )
+            })?
+            .sign(
+                origin_warp,
+                MOCK_HYPERLANE_LOCAL_DOMAIN,
+                self.warp,
+                TokenMessage {
+                    recipient: recipient.address().into(),
+                    amount: amount.into(),
+                    metadata: Default::default(),
+                }
+                .encode(),
+            );
+
+        // Deliver the message to Dango mailbox.
+        self.suite
+            .execute(
+                relayer,
+                self.mailbox,
+                &mailbox::ExecuteMsg::Process {
+                    raw_message,
+                    raw_metadata,
+                },
+                Coins::new(),
+            )
+            .await
+            .result
+            .map_err(|err| anyhow!(err))?;
+
+        // Return the message ID.
+        Ok(message_id)
+    }
+}
